1. RuntimeConfig添加上传附件的文件系统属性配置 2. 上传头像代码改为使用info-core的附件上传代码 3. 将附件上传代码改为使用info-core的附件上传代码

+++ src/main/java/com/vertonur/user/response/form/ResponseForm.java
@@ -111,7 +111,7 @@ public class ResponseForm extends ValidatorForm {
 				.getService()
 				.getDataManagementService(ServiceEnum.RUNTIME_PARAMETER_SERVICE)
 				.getAttachmentConfig();
-		int maxSize = config.getMaxSize();
+		long maxSize = config.getMaxSize();
 		if (upload != null && upload.getFileSize() != 0
 				&& upload.getFileSize() > maxSize) {
 			arg1.setAttribute("fileInvalid", true);
